Adding the level commands
