+++ src/commands/levels/leaderboard.ts
@@ -3,17 +3,20 @@
  * This code is licensed under MIT license (see LICENSE for details)
  */
 
-const { resolveUser, generateLeaderboardCard } = require("../../utils");
-const { MessageAttachment } = require("discord.js");
+import { Message, MessageAttachment } from "discord.js";
+import { BotCommand } from "../../classes/command";
+import { LevelModel } from "../../models/levels";
+import { generateLeaderboardCard } from "../../utils/leaderboard";
+import { resolveUser } from "../../utils/resolveUser";
 
-module.exports = {
-  name: "leaderboard",
-  aliases: ["leader", "rangliste"],
-  usage: ["leaderboard"],
-  help: "Gibt das aktuelle Leaderboard aus.",
-  execute: async (bot, msg) => {
-    const LevelsModel = bot.db.model("Levels");
-    LevelsModel.find({}, async (error, data) => {
+export class LeaderBoardCommand extends BotCommand {
+  name = "Leaderboard";
+  aliases = ["leader", "rangliste"];
+  usage = ["leaderboard"];
+  help = "Gibt das aktuelle Leaderboard aus.";
+
+  async execute(msg: Message): Promise<void> {
+    LevelModel.find({}, async (error, data) => {
       if (data.length === 0) return msg.reply("Aktuell befindet sich noch Niemand auf dem Leaderboard.");
 
       const sorted = data.sort((a, b) => b.xp - a.xp).slice(0, 5);
@@ -22,7 +25,7 @@ module.exports = {
       for (const userData of sorted) {
         const user = resolveUser(msg, userData.userID);
 
-        let rank = sorted.findIndex((entry) => entry.userID === user.user.id) + 1;
+        let rank = sorted.findIndex((entry) => entry.userID === user?.user?.id) + 1;
         if (rank === 0) rank = sorted.length + 1;
         leaderboard.push({ userData: { level: userData.level, xp: userData.xp, rank: rank }, user: user });
       }
@@ -31,5 +34,5 @@ module.exports = {
       const attachment = new MessageAttachment(leaderboardCard.toBuffer(), "card.png");
       msg.channel.send("", attachment);
     });
-  },
-};
+  }
+}
