--- src/utils/roundRect.ts
@@ -1,19 +1,0 @@
-/*
- * Copyright (c) 2020 - 2021 The Pegabot authors
- * This code is licensed under MIT license (see LICENSE for details)
- */
-
-import Canvas from "./Canvas";
-
-export const roundRect = (ctx: Canvas.CanvasRenderingContext2D, x: number, y: number, w: number, h: number, r = 0): Canvas.CanvasRenderingContext2D => {
-  if (w < 2 * r) r = w / 2;
-  if (h < 2 * r) r = h / 2;
-  ctx.beginPath();
-  ctx.moveTo(x + r, y);
-  ctx.arcTo(x + w, y, x + w, y + h, r);
-  ctx.arcTo(x + w, y + h, x, y + h, r);
-  ctx.arcTo(x, y + h, x, y, r);
-  ctx.arcTo(x, y, x + w, y, r);
-  ctx.fill();
-  return ctx;
-};
